# Swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Fundoo Notes Swagger",
        "description": "This is a fundoo notes swagger to access rest API server ",
        "version": "1.0.0",
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:6006",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/file/upload": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "uploads an image",
                "description": "",
                "operationId": "uploadFile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "photos",
                        "in": "formData",
                        "description": "file to upload",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image successful uploaded"
                    }
                }
            }
        },
...
    },
...
}